Rename and move FxManager
Is is actually SteamAudioMapMeshLoader
