+++ src/snd_mem.cpp
@@ -4,7 +4,7 @@
 #include "Loaders/SoundLoader.hpp"
 
 std::shared_ptr<MetaAudio::AudioCache> cache = std::make_shared<MetaAudio::AudioCache>();
-MetaAudio::SoundLoader loader(cache);
+static MetaAudio::SoundLoader loader(cache);
 
 aud_sfxcache_t *S_LoadSound(sfx_t *s, aud_channel_t *ch)
 {
